// src/backend/otel/proto/opentelemetry/proto/collector/metrics/v1/metrics_service.ts
export interface ExportMetricsServiceRequest {
  /**
   * @generated from protobuf field: repeated opentelemetry.proto.metrics.v1.ResourceMetrics resource_metrics = 1
   */
  resourceMetrics: ResourceMetrics[];
}
// ...
class ExportMetricsServiceRequest$Type extends MessageType<ExportMetricsServiceRequest> {
  constructor() {
    super("opentelemetry.proto.collector.metrics.v1.ExportMetricsServiceRequest", [
      { no: 1, name: "resource_metrics", kind: "message", repeat: 2, /*RepeatType.UNPACKED*/ T: () => ResourceMetrics },
    ]);
  }
  override create(value?: PartialMessage<ExportMetricsServiceRequest>): ExportMetricsServiceRequest {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.resourceMetrics = [];
    if (value !== undefined) {
      reflectionMergePartial<ExportMetricsServiceRequest>(this, message, value);
    }
    return message;
  }
  override internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: ExportMetricsServiceRequest,
  ): ExportMetricsServiceRequest {
    let message = target ?? this.create(), end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* repeated opentelemetry.proto.metrics.v1.ResourceMetrics resource_metrics */ 1:
          message.resourceMetrics.push(ResourceMetrics.internalBinaryRead(reader, reader.uint32(), options));
          break;
        default:
          let u = options.readUnknownField;
          if (u === "throw") {
            throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
          }
          let d = reader.skip(wireType);
          if (u !== false) {
            (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
          }
      }
    }
    return message;
  }
  override internalBinaryWrite(
    message: ExportMetricsServiceRequest,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* repeated opentelemetry.proto.metrics.v1.ResourceMetrics resource_metrics = 1; */
    for (let i = 0; i < message.resourceMetrics.length; i++) {
      ResourceMetrics.internalBinaryWrite(
        message.resourceMetrics[i],
        writer.tag(1, WireType.LengthDelimited).fork(),
        options,
      ).join();
    }
    let u = options.writeUnknownFields;
    if (u !== false) {
      (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
    }
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message opentelemetry.proto.collector.metrics.v1.ExportMetricsServiceRequest
 */
export const ExportMetricsServiceRequest = new ExportMetricsServiceRequest$Type();